Initial integration of the NPC Civ system with the player move and gun system in the main scene.

+++ Assets/Gun.cs
@@ -13,7 +13,7 @@ public class Gun : ScriptableObject
     [Header("Bullet Fire")]
     public GameObject bulletType;
     public float bulletSpeed;
-    public float bulletDamage;
+    public int bulletDamage;
     public float bulletForce; //Force multiplier when hitting movable objects
     [Space]
     [Header("Gun Stats")]
